# src/swagger.json
{
	"components": {
...
		"schemas": {
			"CheckInRes": {
				"properties": {
					"card_no": {
						"type": "number",
						"format": "double",
						"description": "Card number"
					},
					"state": {
						"type": "string",
						"description": "The User state after checkin"
					},
					"prev_state": {
						"type": "string",
						"description": "The Previous user state before checkin. The value is 'checkIn' if duplicate checkin operation."
					},
					"notice": {
						"type": "boolean",
						"description": "A discord notification result when an available count is less than 5"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ResJson_CheckInRes_": {
				"properties": {
					"status": {
						"type": "number",
						"format": "double"
					},
					"code": {
						"type": "number",
						"format": "double"
					},
					"result": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"payload": {
						"$ref": "#/components/schemas/CheckInRes"
					}
				},
				"required": [
					"status",
					"code",
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ErrorJson": {
				"properties": {
					"status": {
						"type": "number",
						"format": "double"
					},
					"code": {
						"type": "number",
						"format": "double"
					},
					"result": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					}
				},
				"required": [
					"status",
					"code",
					"result",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResJson_null_": {
				"properties": {
					"status": {
						"type": "number",
						"format": "double"
					},
					"code": {
						"type": "number",
						"format": "double"
					},
					"result": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"payload": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					}
				},
				"required": [
					"status",
					"code",
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ResJson_any_": {
				"properties": {
					"status": {
						"type": "number",
						"format": "double"
					},
					"code": {
						"type": "number",
						"format": "double"
					},
					"result": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"payload": {}
				},
				"required": [
					"status",
					"code",
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"CookieAuth": {
				"type": "apiKey",
				"name": "w_auth_local",
				"in": "cookie"
			},
			"OAuth42": {
				"type": "oauth2",
				"flows": {
					"clientCredentials": {
						"type": "oauth2",
						"authorizationUrl": "http://localhost:3000/user/login",
						"tokenUrl": "http://localhost:3000/user/login/token",
						"refreshUrl": "http://localhost:3000/user/token/refresh",
						"scopes": []
					}
				}
			}
		}
	},
	"info": {
		"title": "42checkin",
		"version": "2.2.7",
		"description": "Cluster checkin service for 42 Seoul cadets.",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "@ryukim, @yurlee, @ohjongin"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/v1/user/checkIn/{cardId}": {
			"post": {
				"operationId": "CheckIn",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResJson_CheckInRes_"
								},
								"examples": {
									"Example 1": {
										"value": {
											"status": 200,
											"result": true,
											"code": 2000,
											"payload": {
												"card_no": 10,
												"state": "checkIn",
												"prev_state": "checkOut",
												"notice": false
											}
										}
									},
									"Example 2": {
										"value": {
											"status": 200,
											"result": false,
											"code": 4090,
											"message": "이미 체크인 상태입니다",
											"payload": {
												"card_no": 10,
												"state": "checkIn",
												"prev_state": "checkIn",
												"notice": false
											}
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorJson"
								},
								"examples": {
									"Example 1": {
										"value": {
											"status": 401,
											"code": 4010,
											"message": "사용자 정보 없음",
											"result": false
										}
									}
								}
							}
						}
					},
					"406": {
						"description": "Not Acceptable",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorJson"
								},
								"examples": {
									"Example 1": {
										"value": {
											"status": 406,
											"code": 4060,
											"message": "체크인 가능 시간이 아닙니다.",
											"result": false
										}
									}
								}
							}
						}
					},
					"409": {
						"description": "Not Acceptable",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorJson"
								},
								"examples": {
									"Example 1": {
										"value": {
											"status": 409,
											"code": 4090,
											"message": "이미 사용 중인 카드입니다. cardOwner: cadet",
											"result": false
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorJson"
								},
								"examples": {
									"Example 1": {
										"value": {
											"status": 500,
											"code": 5000,
											"message": "Internal Server Error",
											"result": false
										}
									}
								}
							}
						}
					}
				},
				"description": "유저 체크인",
				"tags": [
					"User"
				],
				"security": [
					{
						"CookieAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "cardId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/v1/user/checkOut": {
			"post": {
				"operationId": "CheckOut",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResJson_null_"
								},
								"examples": {
									"Example 1": {
										"value": {
											"status": 200,
											"code": 2000,
											"result": true,
											"message": "checkout"
										}
									},
									"Example 2": {
										"value": {
											"status": 200,
											"code": 4090,
											"result": false,
											"message": "이미 체크아웃 하셨습니다."
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorJson"
								},
								"examples": {
									"Example 1": {
										"value": {
											"status": 401,
											"code": 4010,
											"result": false,
											"message": "사용자 정보 없음"
										}
									}
								}
							}
						}
					},
					"406": {
						"description": "Not Acceptable",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorJson"
								},
								"examples": {
									"Example 1": {
										"value": {
											"status": 406,
											"code": 4060,
											"message": "체크아웃 가능 시간이 아닙니다.",
											"result": false
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorJson"
								},
								"examples": {
									"Example 1": {
										"value": {
											"status": 500,
											"code": 5000,
											"message": "Internal Server Error",
											"result": false
										}
									}
								}
							}
						}
					}
				},
				"description": "유저 체크아웃",
				"tags": [
					"User"
				],
				"security": [
					{
						"CookieAuth": []
					}
				],
				"parameters": []
			}
		},
...
		"/v1/log/card/{cardId}": {
			"get": {
				"operationId": "GetCardHistory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResJson_any_"
								},
								"examples": {
									"Example 1": {
										"value": {
											"status": 200,
											"code": 2000,
											"result": true,
											"payload": {
												"list": [
													{
														"_id": 13,
														"login": "chaepark",
														"type": "checkIn",
														"card_no": 25,
														"actor": null,
														"deleted_at": null,
														"updated_at": null,
														"created_at": "2022-02-09T11:46:44.000Z",
														"User": {
															"state": "checkIn",
															"_id": 1,
															"login": "chaepark",
															"card_no": 25,
															"log_id": 13
														}
													},
													{
														"_id": 12,
														"login": "chaepark",
														"type": "checkOut",
														"card_no": 25,
														"actor": null,
														"deleted_at": null,
														"updated_at": null,
														"created_at": "2022-02-08T04:18:03.000Z",
														"User": {
															"_id": 1,
															"state": "checkIn",
															"login": "chaepark",
															"card_no": 25,
															"log_id": 13
														}
													}
												],
												"lastPage": 3
											}
										}
									},
									"Example 2": {
										"value": {
											"status": 200,
											"code": 4040,
											"result": false,
											"message": "card Id에 해당하는 정보가 없습니다.",
											"payload": {
												"list": [],
												"lastPage": 3
											}
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad_Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorJson"
								},
								"examples": {
									"Example 1": {
										"value": {
											"status": 400,
											"code": 4000,
											"message": "쿼리 정보 없음",
											"result": false
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								},
								"examples": {
									"Example 1": {
										"value": "Unauthorized"
									}
								}
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorJson"
								},
								"examples": {
									"Example 1": {
										"value": {
											"status": 403,
											"code": 4030,
											"result": false,
											"message": "접근 권한이 없습니다."
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorJson"
								},
								"examples": {
									"Example 1": {
										"value": {
											"status": 500,
											"code": 5000,
											"message": "Internal Server Error",
											"result": false
										}
									}
								}
							}
						}
					}
				},
				"description": "카드 히스토리 조회",
				"tags": [
					"History"
				],
				"security": [
					{
						"CookieAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "cardId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						},
						"example": "25"
					},
					{
						"in": "query",
						"name": "page",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						},
						"example": "1"
					},
					{
						"in": "query",
						"name": "listSize",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						},
						"example": "2"
					}
				]
			}
		},
		"/v1/log/user/{login}": {
			"get": {
				"operationId": "GetUserHistory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResJson_any_"
								},
								"examples": {
									"Example 1": {
										"value": {
											"status": 200,
											"code": 2000,
											"result": true,
											"payload": {
												"list": [
													{
														"_id": 13,
														"login": "chaepark",
														"type": "checkIn",
														"card_no": 25,
														"actor": null,
														"deleted_at": null,
														"updated_at": null,
														"created_at": "2022-02-09T11:46:44.000Z",
														"User": {
															"state": "checkIn",
															"_id": 1,
															"login": "chaepark",
															"card_no": 25,
															"log_id": 13
														}
													},
													{
														"_id": 12,
														"login": "chaepark",
														"type": "checkOut",
														"card_no": 25,
														"actor": null,
														"deleted_at": null,
														"updated_at": null,
														"created_at": "2022-02-08T04:18:03.000Z",
														"User": {
															"state": "checkIn",
															"_id": 1,
															"login": "chaepark",
															"card_no": 25,
															"log_id": 13
														}
													}
												],
												"lastPage": 7
											}
										}
									},
									"Example 2": {
										"value": {
											"status": 200,
											"code": 4040,
											"result": false,
											"message": "login에 해당하는 정보가 없습니다.",
											"payload": {
												"list": [],
												"lastPage": 3
											}
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad_Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorJson"
								},
								"examples": {
									"Example 1": {
										"value": {
											"status": 400,
											"code": 4000,
											"message": "쿼리 정보 없음",
											"result": false
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								},
								"examples": {
									"Example 1": {
										"value": "Unauthorized"
									}
								}
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorJson"
								},
								"examples": {
									"Example 1": {
										"value": {
											"status": 403,
											"code": 4030,
											"result": false,
											"message": "접근 권한이 없습니다."
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorJson"
								},
								"examples": {
									"Example 1": {
										"value": {
											"status": 500,
											"code": 5000,
											"message": "Internal Server Error",
											"result": false
										}
									}
								}
							}
						}
					}
				},
				"description": "유저 히스토리 조회",
				"tags": [
					"History"
				],
				"security": [
					{
						"CookieAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "login",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "chaepark"
					},
					{
						"in": "query",
						"name": "page",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						},
						"example": "1"
					},
					{
						"in": "query",
						"name": "listSize",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						},
						"example": "2"
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "http://localhost:3000",
			"description": "localhost"
		},
		{
			"url": "https://api.checkin.dev.42seoul.io",
			"description": "development"
		}
	]
}